requirements.txt
langgraph==1.1.2
langchain==0.2.12
tavily-python==0.7.23
langchain-openai==0.2.12
langchain-core==0.2.12
langchain-community==0.2.12
pydantic==2.10.3
python-dotenv==1.0.0
arxiv==2.1.0
wikipedia==1.4.0